refactor: extract component as NavigationButtons
